+++ app.py
@@ -1,5 +1,6 @@
 import streamlit as st
-from decimal import Decimal, InvalidOperation
+from decimal import Decimal
+from utils import _to_decimal, _to_int
 
 # ------------------ Konfiguracja / Config ------------------
 st.set_page_config(
@@ -156,28 +157,10 @@ def load_quick_example() -> None:
         st.session_state[k] = EXAMPLE_QUICK[k]
 
 
-def _to_decimal(value: str) -> Decimal:
-    """Convert user input to ``Decimal``. Returns ``Decimal('0')`` on error."""
-    if value is None or value == "":
-        return Decimal("0")
-    try:
-        return Decimal(value.replace(",", "."))
-    except (InvalidOperation, AttributeError):
-        return Decimal("0")
-
-
 def _entered(key: str) -> bool:
     """Return ``True`` if the user explicitly provided a value for ``key``."""
     val = st.session_state.get(key, "")
     return str(val).strip() != ""
-
-
-def _to_int(value: str) -> int:
-    """Convert user input to ``int``. Returns ``0`` on error."""
-    try:
-        return int(_to_decimal(value))
-    except (InvalidOperation, ValueError):
-        return 0
 
 # ------------------ Funkcje matematyczne -------------------
